feat(inflection): Pluggable inflector, smart-comment overridable (#199)
* Pull down comments on constraints also

* Simplify schema snapshots: one per file

* Move inflection utils to graphile-build

* Move preventEmptyResult

* Allow nested extending

* Allow renaming attributes

* Fix corruption of enums

* Pass constraints to inflectors

* Throttle watching introspection

* Omit logic

* Enable omitting certain fields

* Fix incorrect nulls issue

Previously this query would return lots of `null` because no columns were
explicitly requested

```graphql
{
  allPostEmojis(first: 5) {
    nodes {
      nodeId
    }
  }
}
```

* Excluding columns from create/read/update

* Omitting columns from orderBy

* Enable omitting attributes from condition argument

* Register eXecute permission

* More omit logic

* Extend now supports sub-extending

* Don't require omitting X

* More omits

* Fix name of function 😳

+++ packages/graphile-build-pg/src/plugins/PgMutationUpdateDeletePlugin.js
@@ -2,16 +2,15 @@
 import type { Plugin } from "graphile-build";
 import queryFromResolveData from "../queryFromResolveData";
 import debugFactory from "debug";
-import camelCase from "lodash/camelCase";
-import pluralize from "pluralize";
 import viaTemporaryTable from "./viaTemporaryTable";
+import omit from "../omit";
 
 const debug = debugFactory("graphile-build-pg");
 const base64Decode = str => new Buffer(String(str), "base64").toString("utf8");
 
 export default (async function PgMutationUpdateDeletePlugin(
   builder,
-  { pgInflection: inflection, pgDisableDefaultMutations }
+  { pgDisableDefaultMutations }
 ) {
   if (pgDisableDefaultMutations) {
     return;
@@ -41,7 +40,9 @@ export default (async function PgMutationUpdateDeletePlugin(
           GraphQLID,
         },
         pgColumnFilter,
+        inflection,
       } = build;
+      const { pluralize, singularize, camelCase } = inflection;
       if (!isRootMutation) {
         return fields;
       }
@@ -53,8 +54,12 @@ export default (async function PgMutationUpdateDeletePlugin(
               .filter(table => !!table.namespace)
               .filter(
                 table =>
-                  (mode === "update" && table.isUpdatable) ||
-                  (mode === "delete" && table.isDeletable)
+                  (mode === "update" &&
+                    table.isUpdatable &&
+                    !omit(table, "update")) ||
+                  (mode === "delete" &&
+                    table.isDeletable &&
+                    !omit(table, "delete"))
               )
               .reduce((memo, table) => {
                 const TableType = pgGetGqlTypeByTypeId(table.type.id);
@@ -86,19 +91,13 @@ export default (async function PgMutationUpdateDeletePlugin(
                     const sqlValues = [];
                     const inputData =
                       input[
-                        inflection.patchField(
-                          inflection.tableName(table.name, table.namespace.name)
-                        )
+                        inflection.patchField(inflection.tableFieldName(table))
                       ];
                     introspectionResultsByKind.attribute
                       .filter(attr => attr.classId === table.id)
                       .filter(attr => pgColumnFilter(attr, build, context))
                       .forEach(attr => {
-                        const fieldName = inflection.column(
-                          attr.name,
-                          table.name,
-                          table.namespace.name
-                        );
+                        const fieldName = inflection.column(attr);
                         if (
                           fieldName in
                           inputData /* Because we care about null! */
@@ -189,20 +188,17 @@ export default (async function PgMutationUpdateDeletePlugin(
                         mode === "delete"
                           ? "deletePayloadType"
                           : "updatePayloadType"
-                      ](table.name, table.namespace.name),
+                      ](table),
                       description: `The output of our ${mode} \`${tableTypeName}\` mutation.`,
                       fields: ({
                         recurseDataGeneratorsForField,
                         fieldWithHooks,
                       }) => {
-                        const tableName = inflection.tableName(
-                          table.name,
-                          table.namespace.name
-                        );
+                        const tableName = inflection.tableFieldName(table);
                         recurseDataGeneratorsForField(tableName);
                         // This should really be `-node-id` but for compatibility with PostGraphQL v3 we haven't made that change.
                         const deletedNodeIdFieldName = camelCase(
-                          `deleted-${pluralize.singular(table.name)}-id`
+                          `deleted-${singularize(table.name)}-id`
                         );
                         return Object.assign(
                           {
@@ -283,7 +279,7 @@ export default (async function PgMutationUpdateDeletePlugin(
                       );
                     const fieldName = inflection[
                       mode === "update" ? "updateNode" : "deleteNode"
-                    ](table.name, table.namespace.name);
+                    ](table);
                     const InputType = newWithHooks(
                       GraphQLInputObjectType,
                       {
@@ -292,7 +288,7 @@ export default (async function PgMutationUpdateDeletePlugin(
                           mode === "update"
                             ? "updateNodeInputType"
                             : "deleteNodeInputType"
-                        ](table.name, table.namespace.name),
+                        ](table),
                         fields: Object.assign(
                           {
                             clientMutationId: {
@@ -308,10 +304,7 @@ export default (async function PgMutationUpdateDeletePlugin(
                           mode === "update"
                             ? {
                                 [inflection.patchField(
-                                  inflection.tableName(
-                                    table.name,
-                                    table.namespace.name
-                                  )
+                                  inflection.tableFieldName(table)
                                 )]: {
                                   description: `An object where the defined keys will be set on the \`${tableTypeName}\` being ${mode}d.`,
                                   type: new GraphQLNonNull(TablePatch),
@@ -405,6 +398,9 @@ export default (async function PgMutationUpdateDeletePlugin(
 
                   // Unique
                   uniqueConstraints.forEach(constraint => {
+                    if (omit(constraint, mode)) {
+                      return;
+                    }
                     const keys = constraint.keyAttributeNums.map(
                       num => attributes.filter(attr => attr.num === num)[0]
                     );
@@ -413,14 +409,12 @@ export default (async function PgMutationUpdateDeletePlugin(
                         "Consistency error: could not find an attribute!"
                       );
                     }
-                    const simpleKeys = keys.map(k => ({
-                      column: k.name,
-                      table: k.class.name,
-                      schema: k.class.namespace.name,
-                    }));
+                    if (keys.some(key => omit(key, "read"))) {
+                      return;
+                    }
                     const fieldName = inflection[
                       mode === "update" ? "updateByKeys" : "deleteByKeys"
-                    ](simpleKeys, table.name, table.namespace.name);
+                    ](keys, table, constraint);
                     const InputType = newWithHooks(
                       GraphQLInputObjectType,
                       {
@@ -429,7 +423,7 @@ export default (async function PgMutationUpdateDeletePlugin(
                           mode === "update"
                             ? "updateByKeysInputType"
                             : "deleteByKeysInputType"
-                        ](simpleKeys, table.name, table.namespace.name),
+                        ](keys, table, constraint),
                         fields: Object.assign(
                           {
                             clientMutationId: {
@@ -439,10 +433,7 @@ export default (async function PgMutationUpdateDeletePlugin(
                           mode === "update"
                             ? {
                                 [inflection.patchField(
-                                  inflection.tableName(
-                                    table.name,
-                                    table.namespace.name
-                                  )
+                                  inflection.tableFieldName(table)
                                 )]: {
                                   description: `An object where the defined keys will be set on the \`${tableTypeName}\` being ${mode}d.`,
                                   type: new GraphQLNonNull(TablePatch),
@@ -450,13 +441,7 @@ export default (async function PgMutationUpdateDeletePlugin(
                               }
                             : null,
                           keys.reduce((memo, key) => {
-                            memo[
-                              inflection.column(
-                                key.name,
-                                key.class.name,
-                                key.class.namespace.name
-                              )
-                            ] = {
+                            memo[inflection.column(key)] = {
                               description: key.description,
                               type: new GraphQLNonNull(
                                 pgGetGqlInputTypeByTypeId(key.typeId)
@@ -512,13 +497,7 @@ export default (async function PgMutationUpdateDeletePlugin(
                                     sql.fragment`${sql.identifier(
                                       key.name
                                     )} = ${gql2pg(
-                                      input[
-                                        inflection.column(
-                                          key.name,
-                                          key.class.name,
-                                          key.class.namespace.name
-                                        )
-                                      ],
+                                      input[inflection.column(key)],
                                       key.type
                                     )}`
                                 ),
